server/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
openai==1.79.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
